LayoutTests/imported/w3c/web-platform-tests/permissions-policy/payment-allowed-by-permissions-policy-attribute-redirect-on-load.https.sub-expected.txt

CONSOLE MESSAGE: Feature policy 'Payment' check failed for iframe with origin 'https://www1.web-platform.test:9443' and allow attribute ';payment'.

PASS permissions policy allow="payment" allows same-origin navigation in an iframe.
PASS permissions policy allow="payment" disallows cross-origin navigation in an iframe.
